refactor: 역할에 맞게 컴포넌트 분리 #57
